ciao-launcher, deviceinfo: Add the deviceinfo package
Some of the code in launcher is also duplicated in ciao-down.  This commit
moves this code to a separate package so that it can be used by ciao-down.

The move required some modifications to the overseer and its unit tests.
Interestingly the old tests were not doing what they were supposed to be
doing.  They were supposed to pass the overseer dummy versions of some of
the files in /proc.  These files were passed to the overseer but it was
ignoring them.  This could have led to overseer unit test failures on
some devices.
