+++ ciao-launcher/overseer.go
@@ -17,19 +17,16 @@
 package main
 
 import (
-	"bufio"
 	"container/list"
 	"fmt"
 	"os"
 	"path/filepath"
-	"regexp"
-	"strconv"
 	"sync"
-	"syscall"
 	"time"
 
 	"gopkg.in/yaml.v2"
 
+	"github.com/01org/ciao/deviceinfo"
 	"github.com/01org/ciao/payloads"
 	"github.com/01org/ciao/ssntp"
 	"github.com/golang/glog"
@@ -83,6 +80,31 @@ type ovsStatsStatusCmd struct{}
 
 type ovsRunningState int
 
+type deviceInfo interface {
+	GetLoadAvg() int
+	GetFSInfo(path string) (total, available int)
+	GetOnlineCPUs() int
+	GetMemoryInfo() (total, available int)
+}
+
+type realDeviceInfo struct{}
+
+func (realDeviceInfo) GetLoadAvg() int {
+	return deviceinfo.GetLoadAvg()
+}
+
+func (realDeviceInfo) GetFSInfo(path string) (total, available int) {
+	return deviceinfo.GetFSInfo(path)
+}
+
+func (realDeviceInfo) GetOnlineCPUs() int {
+	return deviceinfo.GetOnlineCPUs()
+}
+
+func (realDeviceInfo) GetMemoryInfo() (total, available int) {
+	return deviceinfo.GetMemoryInfo()
+}
+
 const (
 	ovsPending ovsRunningState = iota
 	ovsRunning
@@ -125,10 +147,8 @@ type overseer struct {
 	diskSpaceAvailable int
 	memoryAvailable    int
 	traceFrames        *list.List
-	memInfo            string
-	stat               string
-	loadavg            string
 	statsInterval      time.Duration
+	di                 deviceInfo
 }
 
 type cnStats struct {
@@ -138,150 +158,6 @@ type cnStats struct {
 	availableDiskMB int
 	load            int
 	cpusOnline      int
-}
-
-var memTotalRegexp *regexp.Regexp
-var memFreeRegexp *regexp.Regexp
-var memActiveFileRegexp *regexp.Regexp
-var memInactiveFileRegexp *regexp.Regexp
-var cpuStatsRegexp *regexp.Regexp
-
-func init() {
-	memTotalRegexp = regexp.MustCompile(`MemTotal:\s+(\d+)`)
-	memFreeRegexp = regexp.MustCompile(`MemFree:\s+(\d+)`)
-	memActiveFileRegexp = regexp.MustCompile(`Active\(file\):\s+(\d+)`)
-	memInactiveFileRegexp = regexp.MustCompile(`Inactive\(file\):\s+(\d+)`)
-	cpuStatsRegexp = regexp.MustCompile(`^cpu[0-9]+.*$`)
-}
-
-func grabInt(re *regexp.Regexp, line string, val *int) bool {
-	matches := re.FindStringSubmatch(line)
-	if matches != nil {
-		parsedNum, err := strconv.Atoi(matches[1])
-		if err == nil {
-			*val = parsedNum
-			return true
-		}
-	}
-	return false
-}
-
-func getMemoryInfo() (total, available int) {
-
-	total = -1
-	available = -1
-	free := -1
-	active := -1
-	inactive := -1
-
-	file, err := os.Open("/proc/meminfo")
-	if err != nil {
-		return
-	}
-	defer func() {
-		_ = file.Close()
-	}()
-
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() && (total == -1 || free == -1 || active == -1 ||
-		inactive == -1) {
-		line := scanner.Text()
-		for _, i := range []struct {
-			v *int
-			r *regexp.Regexp
-		}{
-			{&free, memFreeRegexp},
-			{&total, memTotalRegexp},
-			{&active, memActiveFileRegexp},
-			{&inactive, memInactiveFileRegexp},
-		} {
-			if *i.v == -1 {
-				if grabInt(i.r, line, i.v) {
-					break
-				}
-			}
-		}
-	}
-
-	if free != -1 && active != -1 && inactive != -1 {
-		available = (free + active + inactive) / 1024
-	}
-
-	if total != -1 {
-		total = total / 1024
-	}
-
-	return
-}
-
-func getOnlineCPUs() int {
-
-	file, err := os.Open("/proc/stat")
-	if err != nil {
-		return -1
-	}
-	defer func() {
-		_ = file.Close()
-	}()
-
-	scanner := bufio.NewScanner(file)
-	if !scanner.Scan() {
-		return -1
-	}
-
-	cpusOnline := 0
-	for scanner.Scan() && cpuStatsRegexp.MatchString(scanner.Text()) {
-		cpusOnline++
-	}
-
-	if cpusOnline == 0 {
-		return -1
-	}
-
-	return cpusOnline
-}
-
-func getFSInfo(instancesDir string) (total, available int) {
-
-	total = -1
-	available = -1
-	var buf syscall.Statfs_t
-
-	if syscall.Statfs(instancesDir, &buf) != nil {
-		return
-	}
-
-	if buf.Bsize <= 0 {
-		return
-	}
-
-	total = int((uint64(buf.Bsize) * buf.Blocks) / (1000 * 1000))
-	available = int((uint64(buf.Bsize) * buf.Bavail) / (1000 * 1000))
-
-	return
-}
-
-func getLoadAvg() int {
-	file, err := os.Open("/proc/loadavg")
-	if err != nil {
-		return -1
-	}
-	defer func() {
-		_ = file.Close()
-	}()
-
-	scanner := bufio.NewScanner(file)
-	scanner.Split(bufio.ScanWords)
-	if !scanner.Scan() {
-		return -1
-	}
-
-	loadFloat, err := strconv.ParseFloat(scanner.Text(), 64)
-	if err != nil {
-		return -1
-	}
-
-	return int(loadFloat)
 }
 
 func (ovs *overseer) roomAvailable(cfg *vmConfig) bool {
@@ -485,10 +361,10 @@ func (ovs *overseer) sendTraceReport() {
 func getStats(instancesDir string) *cnStats {
 	var s cnStats
 
-	s.totalMemMB, s.availableMemMB = getMemoryInfo()
-	s.load = getLoadAvg()
-	s.cpusOnline = getOnlineCPUs()
-	s.totalDiskMB, s.availableDiskMB = getFSInfo(instancesDir)
+	s.totalMemMB, s.availableMemMB = deviceinfo.GetMemoryInfo()
+	s.load = deviceinfo.GetLoadAvg()
+	s.cpusOnline = deviceinfo.GetOnlineCPUs()
+	s.totalDiskMB, s.availableDiskMB = deviceinfo.GetFSInfo(instancesDir)
 
 	return &s
 }
@@ -701,8 +577,8 @@ DRAIN:
 	glog.Info("Overseer exitting")
 }
 
-func startOverseerFull(instancesDir string, wg *sync.WaitGroup, ac *agentClient, statsInterval time.Duration,
-	memInfo, stat, loadavg string) chan<- interface{} {
+func startOverseerFull(instancesDir string, wg *sync.WaitGroup, ac *agentClient,
+	statsInterval time.Duration, di deviceInfo) chan<- interface{} {
 
 	instances := make(map[string]*ovsInstanceState)
 	ovsCh := make(chan interface{})
@@ -771,9 +647,7 @@ func startOverseerFull(instancesDir string, wg *sync.WaitGroup, ac *agentClient,
 		memoryAllocated:    memoryAllocated,
 		traceFrames:        list.New(),
 		statsInterval:      statsInterval,
-		memInfo:            memInfo,
-		stat:               stat,
-		loadavg:            loadavg,
+		di:                 di,
 	}
 	ovs.parentWg.Add(1)
 	glog.Info("Starting Overseer")
@@ -798,5 +672,6 @@ func startOverseerFull(instancesDir string, wg *sync.WaitGroup, ac *agentClient,
 }
 
 func startOverseer(wg *sync.WaitGroup, ac *agentClient) chan<- interface{} {
-	return startOverseerFull(instancesDir, wg, ac, time.Second*statsPeriod, "/proc/meminfo", "/proc/stat", "/proc/loadavg")
+	return startOverseerFull(instancesDir, wg, ac, time.Second*statsPeriod,
+		realDeviceInfo{})
 }
