+++ deviceinfo/README.md
@@ -1,0 +1,7 @@
+# deviceinfo
+
+The deviceinfo package provides some utility functions for retrieving
+information about a device, such as how many CPUs it has, or how much
+free space it has available.  These functions are used by both ciao-launcher
+and ciao-down and so are placed in a separate package to maximize code
+re-use.
